analysis_service/requirements.txt
fastapi>=0.68.0
uvicorn>=0.15.0
sqlalchemy>=1.4.0
aiomysql>=0.0.21
ccxt>=3.0.0
pandas>=1.3.0
numpy>=1.21.0
scipy>=1.7.0
python-dotenv>=0.19.0
pytest>=7.0.0
httpx>=0.23.0
greenlet>=3.0.0
apscheduler>=3.10.0
tenacity>=8.2.0
scikit-learn>=1.0.0
